+++ packages/server/src/runtime/builtin/archive.rs
@@ -167,7 +167,7 @@ fn tgar(
 	let server = server.clone();
 	async move {
 		// Create the tgar builder.
-		let mut tgar = tgar::Builder::new(writer).await?;
+		let mut tgar = tgar::write::Builder::new(writer);
 
 		// Write the archive header.
 		tgar.append_archive_header().await?;
@@ -182,7 +182,7 @@ fn tgar(
 async fn tgar_inner<W>(
 	server: &Server,
 	artifact: &tg::Artifact,
-	tgar: &mut tgar::Builder<W>,
+	tgar: &mut tgar::write::Builder<W>,
 ) -> tg::Result<()>
 where
 	W: tokio::io::AsyncWrite + Unpin + Send + Sync,
